converting tests to only read csv files once and pass into json2csv and csv2json test files
