+++ test/tests.js
@@ -1,7 +1,31 @@
 var json2csvTests = require('./testJson2Csv'),
-    csv2jsonTests = require('./testCsv2Json');
+    csv2jsonTests = require('./testCsv2Json'),
+    jsonTestData = require('./testJsonFilesList'),
+    async = require('async'),
+    fs = require('fs'),
+    _ = require('underscore'),
+    csvTestData  = {unQuoted: {}, quoted: {}}, // Document where all csv files will be loaded into
+    csvFiles = require('./testCsvFilesList');
+
+var readCsvFile = function (masterKey, fileInfo, callback) {
+    csvTestData[masterKey][fileInfo.key] = fs.readFileSync(fileInfo.file).toString();
+    return callback(null, csvTestData[masterKey][fileInfo.key]);
+};
 
 describe('json-2-csv Module', function() {
-    json2csvTests.runTests();
-    csv2jsonTests.runTests();
+    before(function(done) {
+        async.parallel(
+            _.flatten(_.map(csvFiles, function (grouping) {
+                return _.map(grouping.files, function (fileInfo) {
+                    return _.partial(readCsvFile, grouping.key, fileInfo);
+                });
+            })),
+            function(err, results) {
+                if (err) throw err;
+                done();
+            });
+    });
+
+    json2csvTests.runTests({jsonTestData: jsonTestData, csvTestData: csvTestData});
+    csv2jsonTests.runTests({jsonTestData: jsonTestData, csvTestData: csvTestData});
 });
